Adding a lot of needed expections.
