+++ .idea/workspace.xml
@@ -1,10 +1,10 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <project version="4">
   <component name="ChangeListManager">
-    <list default="true" id="adc64b2c-c5df-46f2-be04-52434051ae3f" name="Changes" comment="A lot of changes to the structure of the project.">
+    <list default="true" id="adc64b2c-c5df-46f2-be04-52434051ae3f" name="Changes" comment="Adding a lot of needed expections.">
       <change beforePath="$PROJECT_DIR$/.idea/workspace.xml" beforeDir="false" afterPath="$PROJECT_DIR$/.idea/workspace.xml" afterDir="false" />
       <change beforePath="$PROJECT_DIR$/Model/Client.py" beforeDir="false" afterPath="$PROJECT_DIR$/Model/Client.py" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/Model/FIxedHeader.py" beforeDir="false" afterPath="$PROJECT_DIR$/Model/FIxedHeader.py" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/Model/FIxedHeader.py" beforeDir="false" afterPath="$PROJECT_DIR$/Model/FixedHeader.py" afterDir="false" />
       <change beforePath="$PROJECT_DIR$/Model/Server.py" beforeDir="false" afterPath="$PROJECT_DIR$/Model/Server.py" afterDir="false" />
     </list>
     <option name="SHOW_DIALOG" value="false" />
@@ -186,7 +186,14 @@
       <option name="project" value="LOCAL" />
       <updated>1637100238498</updated>
     </task>
-    <option name="localTasksCounter" value="6" />
+    <task id="LOCAL-00006" summary="Adding a lot of needed expections.">
+      <created>1637141835380</created>
+      <option name="number" value="00006" />
+      <option name="presentableId" value="LOCAL-00006" />
+      <option name="project" value="LOCAL" />
+      <updated>1637141835380</updated>
+    </task>
+    <option name="localTasksCounter" value="7" />
     <servers />
   </component>
   <component name="Vcs.Log.Tabs.Properties">
@@ -219,7 +226,8 @@
     <MESSAGE value="Socket interaction is moved on a thread." />
     <MESSAGE value="First attepnts at interptreting a fixed header." />
     <MESSAGE value="A lot of changes to the structure of the project." />
-    <option name="LAST_COMMIT_MESSAGE" value="A lot of changes to the structure of the project." />
+    <MESSAGE value="Adding a lot of needed expections." />
+    <option name="LAST_COMMIT_MESSAGE" value="Adding a lot of needed expections." />
   </component>
   <component name="XDebuggerManager">
     <breakpoint-manager>
@@ -228,6 +236,11 @@
           <url>file://$PROJECT_DIR$/main.py</url>
           <line>27</line>
           <option name="timeStamp" value="8" />
+        </line-breakpoint>
+        <line-breakpoint enabled="true" suspend="THREAD" type="python-line">
+          <url>file://$PROJECT_DIR$/Model/FixedHeader.py</url>
+          <line>31</line>
+          <option name="timeStamp" value="17" />
         </line-breakpoint>
       </breakpoints>
       <default-breakpoints>
@@ -238,10 +251,5 @@
         </breakpoint>
       </default-breakpoints>
     </breakpoint-manager>
-    <watches-manager>
-      <configuration name="PythonConfigurationType">
-        <watch expression="encodedByte[3]" language="Python" />
-      </configuration>
-    </watches-manager>
   </component>
 </project>
