+++ Model/Client.py
@@ -1,4 +1,4 @@
-from Model.FIxedHeader import ProcessFixedHeader
+from Model.FixedHeader import ProcessFixedHeader, RL_Decode
 
 
 class Client:
@@ -10,13 +10,12 @@ class Client:
 
     #This is the soul of our app.
     def ReadMQTTPackage(self):
-        fixedHeader = self.socket.recv(8)
 
-        print(f"The fixed header is: {fixedHeader}")
-        ProcessFixedHeader(fixedHeader)
+        print(f"The fixed header is: {self.socket}") # Reads 8 bites
+        ProcessFixedHeader(self)
 
         #Only needs exatly as much as it needs
-        remainingLenght = RL_Decode(self.socket)
+        remainingLenght = RL_Decode(self)
         print(f" The size of the pachage is:{remainingLenght}")
 
         restOfPachet = self.socket.recv(remainingLenght)
@@ -24,43 +23,4 @@ class Client:
         print(f" The rest of the pachage is:{restOfPachet}")
         return
 
-# def RL_Encode(x):
-#     while True:
-#         encodedByte = x % 128
-#         x = x / 128
-#         if x > 0:
-#             encodedByte = encodedByte | 128
-#         res = encodedByte
-#         if x < 0:
-#             break
-#     return res
 
-
-def RL_Decode(conn):
-    bytesUsed = 1
-    res = 0
-
-    multiplier = 1
-    while True:
-        brah = conn.recv(8);
-
-        encodedByte = to_int(brah)
-        #print(f"\n{encodedByte}")
-
-        res += (encodedByte & 127) * multiplier
-        multiplier *= 128
-        if multiplier > 128 * 128 * 128:
-            raise Exception("Malformed Remaining Length")
-        if encodedByte & 128 == 0:
-            break
-    return res
-
-def to_int(x):
-    enc=x.decode('utf-8')
-    aux=1
-    res=0
-    for i in range(1,7):
-        if enc[i]=='1':
-            res+=aux
-        aux = aux * 2
-    return int(res)
